backend/requirements.txt
Django>=4.2,<5.0
djangorestframework
django-cors-headers
djangorestframework-simplejwt
drf-spectacular
whisper
sentence-transformers
mediapipe
librosa
